vsh_insert_text() re-raise exception if unknown
vsh_insert_text() catches an exception to account for newlines
characters in some of the lines. We should make sure that if that wasn't
the case, and we caught an exception of the same kind with a different
cause that we should re-raise the exception instead of attempting to
insert text again.
While any hypothetical exception would probably have been re-raised by
the second call to vsh_insert_helper() we can't be sure.