Allow dedenting lines in VshSend
This means that sending lines of a python file into a buffer running a
python instance is much easier.
Also make mappings for sending lines of text to a vsh buffer use a
default buffer that can get set once and used from then on.
This means you don't have to keep specifying which buffer to send lines
to.