Operator pending <C-N> shouldn't ever remove above
In operator pending mode, vsh#vsh#MoveToNextPrompt() moves back up one
line so that it leaves the prompt below the end of the output alone.

When it is triggered at the start of a prompt line, this means it moves
to the current prompt (forwards on the same line), and then moves up one
line.
This acts on a line above the current position, which doesn't fit the
idea of operating on lines to the next prompt.

We add a special case so this doesn't happen.