Add special case functions for Python
This programming language is treated specially because of how text can't
be directly pasted from a script file into a REPL.
E.g. empty lines in the middle of a function.

I don't know of any other languages that have this problem.

The special cases add buffer local mappings that remove all blank lines
from the range before sending it to the buffer.