Change definition of command and motion prompts
Until now, prompts with just whitespace after them have been ignored
when running commands.
We now include them.

This means that blank lines can be added, which is nice for things like
giving commands to python (where blank lines tell the REPL that a block
has finished), and passing text into `cat`, where blank lines may be
wanted in the file being created.

I've also removed the special treatment of <Tab>.
Until now this character has been treated as a command character, which
meant that if it were the first character after a prompt
1) Motion would go to before the tab, and not the first word-character
after it.
2) The line would be treated as a command whether or not there was a #
character after the <Tab>

The change in how I'm treating the <Tab> character is mainly because I
don't remember why I treated it that way in the first place, but also
because the only use I can see for the previous behaviour is still
available now that whitespace-only commands are allowed.