Change our wrapper around netrw's `gx` mapping
The previous method just didn't work, and also caused the original `gx`
binding to never be made by netrw (because hasmapto() checks for a
string *anywhere* in an rhs, not just at the start).

Instead of using vsh#vsh#WithPathSet(), we use the same method as for
FileCompletion().

Also reformat the mapping definitions in plugin/vsh.vim.
It's easier to read this way, despite the long lines.