Put some thought into function definition order
Collect functions based on attributes I think it would help for
reading, and ensure that all functions are used after they're defined.

Used after they're defined happened to not matter because I was using a
function right at the bottom of the autoload/ file first, which meant
all the out-of-order usages happened after definitions anyway.