Account for programs with \n only line-endings
There isn't a practical way to tell whether a program is using \n as its
line breaks or \r\n.
Some programs use \n, this means we have to account for both options.
The decision I've made is to remove any trailing \r characters that
occur, but a line split is determined by \n.

This could be made more accurate by looking at recent output to see
whether there was a lot of \r\n or a lot of \n, and guessing based on
that what the line ending sequence is.
I think this would be way too much work for very little gain.