Ensure that plugin is loaded whenever we need it
On recent neovim builds we can no longer be certain that a
plugin/<whatever>.vim file will be sourced before an
ftplugin/<whatever>.vim file.  In fact when opening nvim on a specific
buffer the opposite is always true.

We need the global variables defined in plugin/vsh.vim.  Hence manually
source them.