s:subprocess_closed double-checks before unsetting vsh_job
With recent changes in neovim, starting neovim by opening up a .vsh file
can trigger the filetype setup, then undo the filetype setup, before
starting the filetype setup again.

This is slightly ugly, but it does at least mean that it's exposed a
latent bug in my handling of on_exit callbacks.  I would clear the
`vsh_job` unconditionally when seeing that a process associated with
this buffer has closed.

This was subject to a race condition where if I were to close the
process and open a new one in quick succession, then we could overwrite
the original b:vsh_job variable with that referring to the new process
before the callback associated with the old subprocess was triggered and
cleared that information.

Here we simply check if the b:vsh_job variable is the same as the
a:job_id that this callback was triggered for and avoid clearing the
variable if that is not the case.