s:teardown_mappings be explicit about buffer locality
This function is called to remove mappings for the current buffer when
we are undoing the current filetype plugin.  The commands it removes are
defined to be buffer local, and though everything seems to be working it
seems nice to explicitly show that the commands we're trying to remove
are defined for this current buffer.