Adjust environment variable for Neovim PR #11009
The above PR changed the environment variable that is used to
communicate the Neovim server address to subprocesses.

This commit updates Vsh to use the new variable.  We just try both
possible names for the relevant environment variable rather than check
the neovim version just because it's easier to write.

This commit also adjusts the python package we load in order to access
neovim commands.  Until now it was `neovim` but that's a deprecated
package name and the reference API implementation is `pynvim`.