Fix for vsh#vsh#FileCompletion reverting working directory
I've been noticing a problem for a while where the working directory of
my vim instance would be changed into the directory of the underlying
shell that VSH is running.

It turns out that this is due to the method by which I reset the current
directory after a file-completion command.  This had a problem when
running a second file-completion command directly after the first.

In that situation we would:
  1) Call the FileCompletion command.
     - Save current working directory.
     - Register an autocommand to *undo* the previous save.
     - Return C-x C-f as the keys to insert.
  2) Call FileCompletion command again.
     - Save the *new* working directory (the temporary one).
     - Remove the previous autocommand and register a new one which
       did the same thing.
     - Return C-x C-f as the keys to insert.
  3) Running the base C-x C-f meant that CompleteDone autocommands get
     triggered.
     - Revert working directory to the temporary one.
     - Remove variables saved for which working directory to go back to.

From then on we'd lost the information about what working directory we
were originally in and stayed in the working directory of the underlying
process.

This commit makes two fundamental changes.  First, it stores a
generation number so that the autocommand triggered on CompleteDone can
tell the difference between being triggered when there is another
completion coming straight away and when we're about to drop back into
non-completion mode.  This is the functional change that allows us to
only change our directory back when we're not staying in completion mode
and to ensure that we don't overwrite our recorded "previous directory".

Second, we move the information that we store from the buffer variable
dictionary into a script-local one which maps buffer numbers to recorded
information.  We had originally stored information on the buffer so that
the autocommand had access to that information.  Here the autocommand
calls an autoload vsh function instead which then has access to the
script-local variable.  Given the extra functionality it makes sense to
be using a function anyway.
N.b. This part is just supposed to be an implementation change.