Slight fix and neaten of WithPathSet
I recently noticed that a non-empty string is falsey in vimL.  Expecting
it to behave in the same way as python meant that I had a bug in
WithPathSet.

In that same function I noticed that I implemented a method to reset the
path of that buffer using the BufLeave autocommand.  This seems
unnecessarily complicated given that the function can simply run
whatever command I wanted to run and then unset the relevant buffer
using `setbufvar`.

Possible that I am missing something.
To record the testcases that I used in order to check this (because
writing automated testcases is somewhat tricky):
  - Open a `vsh` file.
  - Use `setlocal` on that file to change the local path.
  - Use one of the WithpathSet mappings (gf, or <CTRL-w>f).
  - Go back to the original buffer.
  - Inspect `&l:path`.
  - The path should be the same as it originally was, with the
    truthy/falsey bug it had reverted to the global path.