cd_to_cwd use `getcwd` unless stored prev_wd is valid
6a5417cab9e18e36d0aa1e4952205fbadec184a1 Fixed the handling of
setting the current directory for FileCompletion so that our "reset"
did not get overwritten by sequential ordering.

However, that commit forgot to ensure that we used a newly-calculated
value for the buffers "proper" directory when called outside of the
recursive state.  This would mean we were essentially using stale
information for the rest of the session.

Test to see this problem:
  - Open a `vsh` buffer.
  - Run file completion on some filename.
  - Change buffer directory with `:cd` in vim.
  - Run file completion on some filename.
  - Observe that the buffer directory has changed back to the original
    one.

This change ensures that the manual buffer directory change has been
preserved across a call to the FileCompletion handler.