Read input from the underlying terminal in raw mode
Many programs run in the terminal do not act linewise.  E.g. prompts
tend to be printed out without a newline after them.

Before this change when running on original vim information sent without
a newline after it would not be sent to our callbacks.  Hence they would
not be echo'd into our file.

This was simply job_start using default values for some arguments,
changing those values to specify that we want the channel to be opened
in `raw` mode solves our problem.

When doing this we then have to change our wrapper function to handle
the new form of data coming in.