Few fixes of using job as number
The original vim version does not like where I check the "truthy" nature
of a job ID since the original vim job ID is a class object.  Checking
that it is not equal to the empty string works though.

Hence changing places which check for `vsh_job` existence to always use
this version of the check.