Enable programs from VSH to command vim
This is naturally available in the neovim implementation, since the
NVIM_LISTEN_ADDRESS is in the environment and neovim naturally waits
around for connections.

Vim does not seem to have a natural "wait for commands" feature other
than +clientserver, and that requires X11 in order to communicate (which
I don't want to rely on).  What vim does have, is the ability to open
sockets *to* something, and receive commands from there.

Here I spawn a python server which provides a "listen" interface to all
processes started in VSH.  This interface waits on an address advertised
by VSH_VIM_LISTEN_ADDRESS, and will receive commands.  As far as the
client to this interface is concerned, replies will look as if coming
from vim.

In order to do this this server also has a "vim" interface to the vim
instance running VSH.  On startup vim opens one channel to this
interface in order to let the server request commands.  When the server
has a new "listen" client it will request vim opens another connection
to handle that client.  Then the server will copy anything this new vim
connection sends to the "listen" client, and will copy anything the
"listen" client sends to the vim connection.

I.e. both vim and the programs we want to spawn wants to behave as a
clients connecting to some known address to talk to each other.  In
order to do that we spawn a process *pretending* to be vim listening on
one address, and *pretending* to be the spawned program listening on
another address.  It then ferries data in between the two.  Puts you in
mind of a reverse-ssh tunnel.