Automatically restore :args after editor finishes
In order to implement $EDITOR, I have a function vsh#vsh#EditFiles.
This sets `args` to the list of files given on the command line.  Until
now the original args were effectively lost.

There was already a function to restore those arguments.  Here we call
that function as soon as the $EDITOR program exits, so that as long as
$EDITOR is not called a second time before the first one exits `args`
will be restored to what it was.

Not perfect, but a great improvement.