Tidyer cleanup around vim exiting
When telling the server to close we also remove our global info.  This
means that other buffers will recognise the server has closed and could
start one again if requested.

We also catch an exception if the `atexit` use of `os.kill` happens on a
closed process.  Without this catch vim would flash an error message on
closing if the VSH server process was already closed.