Allow python server to always be started on StartSubprocess
This means that if we've previously closed the server for some reason,
then when opening *other* vsh files we'll restart it.  You can also
manually trigger it using StartSubprocess, though that will likely
complain that you already have a vsh_job in the current buffer.

While testing that out noticed that the StartSubprocess function uses
`get(b:, 'vsh_job', 0)` and hence trips up in vim when we already have a
job in this buffer.  Replace that with the newly-canonical `getbufvar`
and check for empty string.

Also introduec a new function which closes all existing channels when
the server process is stopped.  Without that server process those
channels will not be receiving anything anyway.  We also take this
opportunity to ensure we clear up the global server-related vim
variables to make it clear what's going on.

Finally this commit also adjusts vsh#vsh#RestoreArgs so it does not
change the current buffer that you are viewing.  This is more important
now that it is run automatically when the VSH $EDITOR is closed.