Add prefix to functions defined in origvim_server_setup
This runs in the vim global namespace (as needed in order to be able to
change variables inside vim).  Hence I feel that defining functions
which do not have a `vsh` prefix seems to be bad-form.

Similarly, in order to avoid polluting the global namespace, I delete
the variables created before returning to vim.