Tidy-ups around $EDITOR
Use one python script for both original vim and neovim.
Cleanly exit but with an exit code when using KeyboardInterrupt (rather
than letting the exception raise).

Noticed that vsh#vsh#RestoreArgs() actually re-edits files, which would
lose state in any .vsh file that has this done.  It also raises an error
when called if the relevant file getting :edit'ed  is modified.

Use `argdelete` and `argadd` to change the argument list without such
`:edit`, but could not find any way to *select* an item on the argument
list without editing it so that is not restotred (for now).