vsh#vsh#RestoreArgs now usually handles argid
Before this commit we would restore the argument list, but we would not
restore which item in the argument list is "current".

This was because most of the commands for changing the argument list
also essentially ran :edit on a file.  That command re-reads a file, and
if done on a `vsh` buffer it essentially restarts the underlying
process (losing all state).  That is not worth keeping the argument list
as expected.

Here I found that choosing how we run `argadd` can determine which id
the current window has selected.  While working on this I noticed that
the "current argument" is often different between different windows.
The mechanism of restoring arguments restores the argument ID that was
in the window where EditFiles was called to the "current" window.

This is not perfect, but since windows can be closed and re-opened
between these two function calls there doesn't seem to be a "perfect"
action to perform.  Essentially we don't know what the `argidx()` should
be for this window since we don't even know if this window is logically
the same as one that was around when the EditFiles was called.

The current behaviour is considered "good enough" :-)