Remove global variables from buffer unlet code
I have a function to remove buffer local variables which is called when
the filetype of VSH is dropped.  It erroneously contains vsh_insert_mark
and vsh_prompt_mark.  These variables are not buffer-local, so they
should not be in this list.

It shouldn't have been causing any problems, but it seems worth removing
for code-clarity purposes.