Directly use `/bin/bash` as shebang for vsh_shell_start
Also add a comment describing thoughts around whether to use an
interactive shell to run `vsh_shell_start`.

N.b. see github issue #32 for context.

Idea behind that is to enable choosing a different INPUTRC in a users
bashrc, and having that propogate to the vsh_shell_start script.
This would allow the `bind -q` queries to pick up changed mappings.

I've seen someone attempt to do just that and figured it would be nice
to make another approach "just work".

Turns out that running the script as an interactive shell has problems
when starting an interactive shell prints something out.  That currently
is happening on my local machine (which is how I spotted this).

The text ends up printed out multiple times into the relevant buffer
(and printed out without the `stty` settings, so including a
carriage-return).  On top of that if you avoid actually loading the
buffer when accidentally starting a new `neovim` instance on the same
VSH file you get a python stack trace printed out onto the terminal.

The trade-off in order to allow INPUTRC being set in bashrc does not
seem worth it (at least for the moment -- until I see this happening
in the wild more often).