Pass VSH_VIM_BUFFER in a jobs environment
Before this was passed to vsh_shell_start as a command line argument.
We looked into changing it since that seemed unnecessary (since we
already have an environment dictionary that we're setting from
vimscript).

Not only does it make things a little cleaner, but since we're now
running vsh_shell_start as a login script it means that the *first* time
that bashrc is sourced it will have this VSH_VIM_BUFFER variable set
(and could determine configuration based on that).

Not that we encourage doing such a thing -- it's recommended to make
changes based on TERM=dumb, since that variable usually gets carried
across into SSH sessions.