Add variables to override vsh mappings
Was quite easy to implement since I have this "mappings defined in a
list of lists" approach.

Want to allow any user to override mappings rather than having to hook
in something to unmap something and then map it to something else.
Approach is to provide global variables which can be assigned to a
mapping string, these override the default mapping at the time that we
create said mapping.

N.b. in this we also switch the mapping to run `possible-completions` in
the underlying shell from <C-q> to <TAB>.  This seems like what would be
most expected to any new user -- see #32.

N.b. if we open a VSH file as the first buffer, then the "run ftplugin
once, undo ftplugin, run again" behaviour that *neovim* only exhibits
gives us a bit of trouble.

If the user sets these `g:` override variables in the `plugin/`
directory then they may not be loaded before the `ftplugin/` commands
run and initiate mappings.  Since the `ftplugin/` stuff gets re-run we
still end up with the mapping that the user requested, but the original
mappings are not overridden but instead both exist.

This happens whenever the `g:` override variable is set *after* having
run `syntax on` in `vimrc`.  As far as neovim is concerned `syntax on`
is not supposed to be in the initialisation file.  Hence this *should*
not be a concern.

We have the same problem if `<mapleader>` and `<maplocalleader>` are
specified after `syntax on` in ones initialisation file.

See the below for discussion on this behaviour
https://github.com/neovim/neovim/issues/19008.