Adjust terminal width in recommended GDB config for VSH
Before I was recommending a width of 0 (which means no wrapping at all).
While this works nicely for things like backtrace and the breakpoint
information lines it makes the output from `possible-completions` much
harder to read (everything is on one line).

Hence change the recommendation to something that is OK for everything.