Introduce vsh#vsh#RestartSubprocess()
A few times I've hit the problem where I was SSH'd into another machine
when the connection dropped, and the shell started hanging.  This commit
adds a function to essentially close the terminal and open a new one.
That makes it easier to handle such a case without having to wait for
the SSH process to time out while waiting for the other side to respond.