vsh_closing_jobs avoid overriding
This variable is there to store which vsh jobs are associated with
unloaded buffers and will hence need to be stopped when said buffers are
deleted.

The ftplugin/vsh.vim script unconditionally sets this to an empty
dictionary when starting.  This creates a race condition between the
unload and deletion of a buffer where if another vsh buffer is loaded we
lose our cached information.

There is no reason to always initialise this variable.  This commit only
initialises it if it does not already exist.