First steps towards including emacs stuff
Including emacs code in order to keep definition of VSH file in one
place.  Hopefully this will mean when I notice a problem with one I
fix it in both.

(Though so far I've noticed differences that I've not yet fixed quite
often, so don't know how much of that is actually going to happen).