Twisting around outer and inner loop
Noticing that the loop does a heck of a lot of changing the text in a
buffer.  I can reduce the amount of this easily by testing with the
same text at each different point in the randomly chosen set of
positions as the *inner* loop.

Before I had the outer loop moving point and the inner loop adjusting
the buffer.