Fix vsh--command-block-bounds
When marking the bounds of a command block if there was no command
block above the current point I don't want to move the cursor.

I happened to be using `point' as the position to return for "end of
command block".  This worked fine except for the fact that I used
`beginning-of-line' to move to the start of a line before checking for
split or command regexps.

This change simply records the *original* point that this function was
called with and returns that when wanting to not move the cursor
instead of the point after `beginning-of-line'.