Random bug-fix for `vsh-new-prompt'
When the last line in a buffer is not terminated with a newline we
would previously insert the new prompt at the end of that line.  This
would not actually generate a command line (because the prompt does
not start the current line).

We fix this by adding an if-clause.  Simply insert a newline if we are
not at the start of a line.