Convert vsh--with-standard-blocks test from macros to functions
I wanted to separate the specific functions from the outling code of
"generate a basic block and call tests with points at different
places".

Originally I did it with a macro in order to introduce certain
variables into the test lexical scope.  I eventually decided that this
was not giving enough benefit for the slight extra confusion, and have
now switched to just passing in some functions to call with all the
extra variables that I wanted to define.