Fix up `vsh--beginning-of-block-fn'
As yet not tidied up the testsuite enough to commit, but while writing
that I found some bugs in this function and updated it.

N.b. this function is now starting to behave in the general approach
of "find next `beginning' point and move to it".  That is making
testing this (and testing other similarly defined motions) much
easier.