Rewriting block motion elisp tests
With a different approach to the definition of how motion commands
work we have a much nicer approach to the tests.

Rather than things like "move to the next start of block
position *unless* you're before the end of prompt on the first line of
a current block" (i.e. linewise motion as per vim), we now define
motion as "to the next start of block point".

This new definition both "feels" better in an emacs context (at least
to me), and is *much* easier to test.