Fix some of the easier compilation warnings
These were fixed with a search/replace.  A bunch of unused lexical
variable warnings may get fixed by taking another approach for the
testing of block marking (something similar to the motion approach I
used).  Will have to figure out what the new way to define a package
is given that `define-package' is an obsolete function.