Adjust meaning of `vsh--end-of-block-fn'
Originally I had this moving to the end of the line on the last line
of the block.  I later realised this meant removing the region
highlighted by `mark-defun' would leave an empty line in between the
outputs above and below.

I figured that extending the definition of the "end of function" to
include the newline at the end of a block would mean this behaviour
would be nicer, plus it would match the behaviour of
`vsh-mark-command-block' with a prefix argument.

Updated tests and functionality.