Improve the filename completion
Still working on this, but it's a process.  It does look like it would
be better to use `comint-dynamic-complete-filename' (with all its
cleverness around environment variables and tildes etc).  However I
would like to use the environment variables of the underlying shell
rather than emacs itself.

I expect I can create a macro like `vsh-with-temp-environment' for
this, but that will have to wait until I have a bit more time.