Fix problem with ansi-color-{filter,apply-on}-region
Was getting some errors in the ansi colouring where emacs would
complain about an invalid search bounds inside the colouring code.
This turned out to be because of some state saved between calls to the
ansi colouring code.

That state is useful when "continuing" on a coloured region, but when
starting a new region it is not useful.

In this commit we check for "continuing" on a coloured region by
whether the "last" insert that triggered colouring ended at the point
where we are just about to insert some new text.

This is not perfect -- there could be text edits in the middle that
make things a little fuzzy -- but as a whole vsh.el treats all
"outputs" as just text coming in at a given point in a buffer and
doesn't have any way to tell whether text coming in from the output of
a command is "finished" or not.
I.e. I don't know of any better way -- though not looked too hard --
and it doesn't feel like there should be a better way, so going with
this.