Ensure we kill the underlying vsh process on changing modes
Without this change I would often end up with a stray `vsh-proc'
process in some info buffers.  This would happen because the `info'
command first creates a new buffer with `pop-to-buffer-same-window',
then changes the mode to `Info-mode'.  The new buffer starts out in
the same mode as the original buffer one started from, which can be
`vsh-mode', and if that is the case we would naturally start the
underlying vsh process.

This change ensures that whenever using `kill-all-local-variables' we
also kill the vsh process for this buffer.  That way such a call would
logically "fully clean" the buffer of vsh-mode state.