Automatically add a new prompt to empty file
I always get a little bothered when starting a new file that I have to
type <localleader>n in order to get started, and that leaves an empty
line at the start of the file.

In emacs my logical implementation is slightly different, such that the
`new-prompt` action does not leave that empty line.

I attempted to avoid that empty line with the same approach that I used
in emacs, but because of the difference between the two implementations
in how I recognise I need to add a newline because this is the logical
"start of input" that approach has the problem that later output would
go *after* this prompt if we insert it at the point where text was due
to be put.

I do not want to standardise how we make this decision between both
emacs and vim (honestly not sure which trade-off is best, quite like
having the difference and watching how things work). Instead I recognise
that the only time I was getting bothered was when opening an empty
file, and solve *that* problem by automatically adding a prompt when
opening an empty file.
This also means that someone trying my plugin for the first time gets
some sort of confirmation that the plugin is doing something when
opening their first file (which is nice).