Fix adaptive-fill setup
When defining `adaptive-fill-first-line-regexp' we originally avoided
changing the value if it was already a `local-variable-p'.  While
attempting to use this setup it seems this means avoiding setting
things.

Honestly not understood this fully -- I never thought that check would
prevent real problems and it seems to be preventing the feature from
working right now, so removing the check.

It could be something like `setq-local' makes the variable buffer
local before running the expression to set it to (in which case an
alternative fix could be to manually evaluate the expression first and
use the result in the `setq-local').

Either way -- whether there's a fix that allows keeping the check for
buffer-local around or not -- since I don't really care about this
check I'm just removing it.

... We'll see if I end up re-enabling it later when I find out that
check would actually prevent some real problems ;-)