Update key mappings
Use `<remap>' a lot more on the belief that the user would prefer
commands be easy and relevant to this major mode than have all
commands around.

Also attempt to move some keys away from the C-c prefix, but failed to
think of where else they should go.  I expect I should parametrise it,
but I still do sort of want the default to be C-c in order to follow
`comint' mappings.