Change `vsh-mark-command-block' to act more like `mark-defun'
This means both that the region expands on multiple calls and that the
point ends up at the start of the block.

Still not matching entirely since we don't take a count but instead
have an argument choosing whether to include comments or not as the
logical "block".  If the user would prefer to have the "count"
functionality than this they could just use `mark-defun'.

Along with this change we also make `vsh-execute-block' use any
currently active region.  That saves having to maintain two
keybindings for very similar functionality.

Also update tests for `vsh-mark-command-block' change.
Since I updated the behaviour of this command I need to update the
test too.  It is somewhat nice that this gives me the excuse to change
from the old "macro and manually specify what to do" approach to using
the "oracle" approach for `vsh-mark-command-block' testing.