Use C-M-x directly for `vsh-execute-block'
I would like to use `<remap> <eval-defun>', but it seems that even
though `eval-last-sexp' is on the global map, `eval-defun' is in the
emacs-lisp and lisp-interaction-mode maps.  Hence the remap thing
doesn't work.

Using the default position for `eval-defun' and expecting anyone that
moved it to also move the mapping for `vsh-execute-block' in the vsh
mode map.