Handle edge-case around `vsh--delete-and-send'
I found that if the region that we're deleting is empty then we don't
deactivate the mark.  If the region we're deleting contains something
then we do.  This means the observed behaviour when running a command
with `vsh-execute-region' could either be to see the region get
deleted or to see the region now cover the output that comes back from
commands.

This commit ensures we always deactivate the region in
`vsh--delete-and-send' in order to provide a more uniform experience.