Fix quirks around alternate name server start
Until this fix, when we used a fallback server name we ended up
without syntax highlighting.  This was unfortunate but I didn't know
what was going wrong.

On debugging it turns out tha the problem was to do with the warning
being raised about an existing server.  This warning meant that the
warning buffer may be created.  When creating that warning buffer
`global-font-lock-mode-enable-in-buffers' can get called.

When this warning is raised the `vsh-mode' buffer has been created but
not fully initialised (most notably the `font-lock-defaults'
information has not been set up yet).  That means that
`font-lock-mode' gets turned on in the buffer but does nothing.

After the warning has been handled and displayed to the user, emacs
returns to finishing off the initialisation of the `vsh-mode' buffer.
This includes setting `font-lock-defaults' and running
`after-major-mode-hook'.  In that hook we again call
`global-font-lock-mode-enable-in-buffers', but by this time
`font-lock-mode' has been turned on in our buffer so it doesn't bother
looking at our buffer a second time.

Hence the `font-lock-mode' setup is run once on our buffer when its
data has not been set, and doesn't bother running a second time after
its data *has* been set.

There are a few ways of fixing this -- one could be to manually handle
`font-lock-mode' setting in the hook that sets `font-lock-defaults'.
Another is to merge some of the hooks into the `vsh-mode' body rather
than having the majority of the setup in hooks.

I chose to separate the creation of an emacs server into yet another
hook and put it in the default `vsh-mode-hook' after the color setup.

To be honest I don't really see much of a difference between these
approaches, and I think the one I chose "looks" neater".