C-u C-a now moves to very beginning of line
I usually want to move to the beginning of the command, but when I
want to move to the very start of the line it's particularly
frustrating if I don't have an easy way to do that.

Hence add the ability to go there with universal argument passed to
`vsh-bol'.