Use #' instead of ' for various references to functions
I'm not actually sure I like this.  Putting a symbol in a hook seems a
bit nicer than putting the function in the hook directly because one
could change the function under than symbol and get everything
updated.  Similar for setting local function names like
`fill-forward-paragraph-function'.

This behaviour only really seems useful for development, so I don't
think it's anything much to be concerned about.  Hence following the
recommendation even if I am not really in favour of the change.