Avoid DeprecationWarning about os.fork
Python 3.12 has added a DeprecationWarning when using `os.fork` in a
multithreaded environment.  This is in general something that needs to
be watched in a POSIX world.

Now since I used `fork` just to `exec` straight after, I thought that I
would be fine.  Moreover, since vim is logically single threaded I
thought I would be fine.

It turns out that (1) The CPython runtime itself is not safe to run
after a fork in a multithreaded environment, and (2) vim is
multithreaded after startup.

I don't fully understand why vim is multithreaded and what the different
threads do, but these two things combined meant that I hit the
DeprecationWarning.

Here we simply use subprocess.Popen instead, that avoids any need for
steps in between the `fork` and `exec` in python itself and is also a
bit cleaner.