Add INPUTRC munging for vim
Much of the interface requires some inputrc stuff.  Until now I relied
on the user setting these things.  This is a bit much to ask for someone
just trying things out (and the `demo` directory I set up to introduce
people to `vsh` for vim is a pain to adjust in order to make it work for
emacs too).

Hence try and munge INPUTRC in order to avoid some of the more
problematic options without overriding too much of the users
configuration.

Hoping this is going to work well -- will have to see how things go.
N.b. Need to add emacs support and link in the new file to the emacs
directory before this goes upstream.