Switch how INPUTRC handling done for emacs
Will be doing the same for vim next.

Change is that instead of the shell determining which file to store a
temporary INPUTRC in, the editor determines this and passes it in.
This is thought better because the bookkeeping for eventually removing
the INPUTRC temporary file is done in the editor and not in the shell,
so keeping the generation of the file in the editor at the same time
avoids some mis-matches in edge-cases (e.g. if `vsh_shell_start' is
run manually by the user).