Again adjust how to record & remove INPUTRC file
Instead of recording it on a per-buffer basis, record it on a
per-process basis.  Then remove it when the buffer closes.

This means we get to use the in-built editor "run on process closed"
hook, avoiding re-writing the functionality of hooking into every way
in which a process could be closed.

Cuts out a good amount of code, avoids having another buffer local
variables, and I expect makes things more robust.
N.b. Still have to hook into kill-emacs-hook since it looks like the
sentinel is not called for that.