Add INPUTRC tempfile handling for original vim
This was a little tricky because correlating the job to a given inputrc
can't be done in the obvious way because I've not found a way to use a
vimL job object as a key in a dictionary.

N.b. may have stumbled across another original vim bug around this --
but I'm not putting much effort into handling original vim (beyond
making it viable enough that people will use it enough to give me bug
reports).